Refs 9671. Consider data sharing pool only if it comes from listener
If the writer is data sharing compatible, it will use a
WriterPool, which is subclassed from DataSharingPayloadPool

If the reader and writer are not compatible between each other, but
are on the same process, the writer will call the reader's
ProcessDataMessage with a CacheChange that contains a payload from the
WriterPool.

The reader was trying to convert the payload's pool to
DataSharingPayloadPool to check if it should be treated as data sharing.
Since WriterPool IS a DataSharingPayloadPool, it would succeed, causing
the reader to mingle with the writer's pool / history.

The reader must check specifically whether the pool is a ReaderPool,
which is the exact type that the listener is using.

Signed-off-by: Iker Luengo <ikerluengo@eprosima.com>