Refs 9671. Avoid read and write race condition
When the listener is notified, it traverses the available payloads
in the pool processing all the available that have not been read yet.
However, the write pointer was advanced as soon as the writer
asked for a payload, even though it had no data yet.
So the listener could try to process a payload that was not ready to be
processed.

Separating the 'next-free-payload' and 'last-ready-to-process-payload'
references solves the issue. The listener will process until the last
ready payload and the writer will advance this reference when all the
data is ready (right before the notification).

Signed-off-by: Iker Luengo <ikerluengo@eprosima.com>